Refactor: create plan::tasks submodule
